Implementado novo Service TransferenciaService.java
