+++ pom.xml
@@ -68,6 +68,36 @@
 					</excludes>
 				</configuration>
 			</plugin>
+			<plugin>
+				<groupId>org.jacoco</groupId>
+				<artifactId>jacoco-maven-plugin</artifactId>
+				<version>0.8.8</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>prepare-agent</goal>
+						</goals>
+					</execution>
+					<execution>
+						<id>report</id>
+						<phase>verify</phase>
+						<goals>
+							<goal>report</goal>
+						</goals>
+					</execution>
+				</executions>
+				<configuration>
+					<excludes>
+						<exclude>**/*Application*</exclude>
+						<exclude>**/entity/*</exclude>
+						<exclude>**/dto/*</exclude>
+						<exclude>**/model/*</exclude>
+						<exclude>**/repository/*</exclude>
+						<exclude>**/security/*</exclude>
+						<exclude>**/config/**</exclude>
+					</excludes>
+				</configuration>
+			</plugin>
 		</plugins>
 	</build>
 </project>
