+++ src/main/java/br/com/banco/controller/AuthController.java
@@ -3,6 +3,7 @@ package br.com.banco.controller;
 import br.com.banco.dto.UsuarioDTO;
 import br.com.banco.entity.ContaBancaria;
 import br.com.banco.entity.Usuario;
+import br.com.banco.service.TransferenciaService;
 import org.springframework.security.core.Authentication;
 import br.com.banco.service.ContaBancariaService;
 import br.com.banco.service.UsuarioService;
@@ -22,10 +23,12 @@ public class AuthController {
 
     private final UsuarioService usuarioService;
     private final ContaBancariaService contaBancariaService;
+    private final TransferenciaService transferenciaService;
 
-    public AuthController(UsuarioService usuarioService, ContaBancariaService contaBancariaService) {
+    public AuthController(UsuarioService usuarioService, ContaBancariaService contaBancariaService, TransferenciaService transferenciaService) {
         this.usuarioService = usuarioService;
         this.contaBancariaService = contaBancariaService;
+        this.transferenciaService = transferenciaService;
     }
 
     @GetMapping("index")
@@ -38,12 +41,12 @@ public class AuthController {
         return "login";
     }
 
-    @GetMapping("register")
-    public String showRegistrationForm(Model model){
-        UsuarioDTO usuarioDTO = new UsuarioDTO();
-        model.addAttribute("user", usuarioDTO);
-        return "register_";
-    }
+//    @GetMapping("register")
+//    public String showRegistrationForm(Model model){
+//        UsuarioDTO usuarioDTO = new UsuarioDTO();
+//        model.addAttribute("user", usuarioDTO);
+//        return "register_";
+//    }
 
     @GetMapping("/usuario/dashboard")
     public String dashboardUsuario(Authentication authentication, Model model) {
@@ -78,6 +81,25 @@ public class AuthController {
             model.addAttribute("titular", conta.getTitular());
             model.addAttribute("email", emailUsuarioLogado);
             model.addAttribute("numeroConta", numeroConta);
+            return "usuario/dashboard";
+        }
+    }
+
+    @PostMapping("/usuario/transferir")
+    public String transferir(@RequestParam String numeroContaDestino, @RequestParam BigDecimal valor,
+                             Authentication authentication, RedirectAttributes redirectAttributes, Model model) {
+        String emailUsuarioLogado = authentication.getName();
+        try {
+            ContaBancaria contaOrigem = contaBancariaService.consultarEmail(emailUsuarioLogado);
+            transferenciaService.transferir(contaOrigem.getNumeroConta(), numeroContaDestino, valor);
+            redirectAttributes.addFlashAttribute("success", "Transferência realizada com sucesso.");
+            return "redirect:/usuario/dashboard";
+        } catch (IllegalArgumentException | IllegalStateException e) {
+            model.addAttribute("error", e.getMessage());
+            ContaBancaria conta = contaBancariaService.consultarEmail(emailUsuarioLogado);
+            model.addAttribute("titular", conta.getTitular());
+            model.addAttribute("email", emailUsuarioLogado);
+            model.addAttribute("numeroConta", conta.getNumeroConta());
             return "usuario/dashboard";
         }
     }
