+++ src/main/java/br/com/banco/repository/ContaBancariaRepository.java
@@ -19,5 +19,4 @@ public interface ContaBancariaRepository extends JpaRepository<ContaBancaria, Lo
 //    @Lock(LockModeType.PESSIMISTIC_WRITE)
 //    @Query("SELECT c FROM ContaBancaria c WHERE c.numeroConta = :numeroConta")
 //    Optional<ContaBancaria> findByNumeroContaWithLock(@Param("numeroConta") String numeroConta);
-
 }
