+++ src/main/java/br/com/banco/service/TransferenciaService.java
@@ -1,0 +1,7 @@
+package br.com.banco.service;
+
+import java.math.BigDecimal;
+
+public interface TransferenciaService {
+    void transferir(String contaOrigem, String contaDestino, BigDecimal valor) throws IllegalArgumentException;
+}
