+++ src/main/resources/templates/usuario/dashboard.html
@@ -46,11 +46,73 @@
         <input type="hidden" id="tipoOperacao" name="tipoOperacao" />
         <div class="form-group">
             <label for="valorOperacao">Valor</label>
-            <input type="number" class="form-control" id="valorOperacao" name="valor" placeholder="Informe o valor" step="0.01" required>
+            <input type="text" class="form-control" id="valorOperacao" name="valor" placeholder="Informe o valor" step="0.01" required
+                   pattern="[0-9]+(\.[0-9]{1,2})?" title="Apenas números e ponto são permitidos">
         </div>
         <button type="submit" class="btn btn-success" onclick="document.getElementById('tipoOperacao').value='CREDITO'">Creditar</button>
         <button type="submit" class="btn btn-danger" onclick="document.getElementById('tipoOperacao').value='DEBITO'">Debitar</button>
     </form>
+
+    <!-- Checkbox para habilitar transferência -->
+    <div class="form-check mb-3">
+        <input class="form-check-input" type="checkbox" id="habilitarTransferencia">
+        <label class="form-check-label" for="habilitarTransferencia">
+            Habilitar Transferência
+        </label>
+    </div>
+
+    <!-- Formulário de transferência -->
+    <div id="transferenciaForm" class="card mb-4 d-none">
+        <div class="card-body">
+            <h5 class="card-title">Transferência</h5>
+            <form method="post" th:action="@{/usuario/transferir}">
+                <div class="form-group mb-3">
+                    <label for="numeroContaDestino">Número da Conta Destino</label>
+                    <input type="text" class="form-control" id="numeroContaDestino" name="numeroContaDestino" placeholder="Informe o número da conta destino" required
+                           pattern="[0-9]+(\.[0-9]{1,2})?" title="Apenas números e ponto são permitidos">
+                </div>
+                <div class="form-group mb-3">
+                    <label for="valorTransferencia">Valor</label>
+                    <input type="text" class="form-control" id="valorTransferencia" name="valor" placeholder="Informe o valor" step="0.01" required
+                           pattern="[0-9]+(\.[0-9]{1,2})?" title="Apenas números e ponto são permitidos">
+                </div>
+                <button type="submit" class="btn btn-primary">Transferir</button>
+            </form>
+        </div>
+    </div>
 </div>
+<script>
+    // Função genérica para permitir apenas números e um único ponto
+    function aplicarMascaraNumerica(inputId) {
+        const input = document.getElementById(inputId);
+        input.addEventListener('input', () => {
+            let valor = input.value.replace(/[^0-9.]/g, '');
+            const partes = valor.split('.');
+            if (partes.length > 2) {
+                valor = partes[0] + '.' + partes.slice(1).join('');
+            }
+            input.value = valor;
+        });
+    }
+
+    // Função para alternar a exibição do formulário de transferência
+    function configurarToggleTransferencia(checkboxId, targetId) {
+        const checkbox = document.getElementById(checkboxId);
+        const target = document.getElementById(targetId);
+
+        checkbox.addEventListener('change', () => {
+            target.classList.toggle('d-none', !checkbox.checked);
+        });
+    }
+
+    // Inicialização dos comportamentos ao carregar a página
+    document.addEventListener('DOMContentLoaded', () => {
+        configurarToggleTransferencia('habilitarTransferencia', 'transferenciaForm');
+
+        aplicarMascaraNumerica('numeroContaDestino');
+        aplicarMascaraNumerica('valorOperacao');
+        aplicarMascaraNumerica('valorTransferencia');
+    });
+</script>
 </body>
 </html>
